Refactor core structure and move files to feature folders
Reorganized the Sources/Scout/Core directory by moving files into feature-specific subfolders (Activity, Helpers, Log, Matrix, Monitor, Setup, Storage, Sync, Telemetry). Added new Syncable conformance files for UserActivity, EventObject, SessionObject, and MetricsObject. Removed redundant files now located in Helpers and Sync folders. This improves codebase modularity and maintainability.
