+++ Sources/Scout/Core/Activity/ActivityPeriod.swift
@@ -7,13 +7,6 @@
 
 import Foundation
 
-/// Represents the period of time for which active users are tracked.
-///
-/// The active user period can be daily, weekly, or monthly. This can be used to
-/// track the number of active users over a specific period of time. For example,
-/// you can track the number of daily active users, weekly active users, or monthly
-/// active users.
-///
 enum ActivityPeriod: String, Identifiable, CaseIterable {
     case daily = "d"
     case weekly = "w"
@@ -32,11 +25,6 @@ enum ActivityPeriod: String, Identifiable, CaseIterable {
         }
     }
 
-    /// A property that returns the appropriate count field key path for the activity period.
-    ///
-    /// This property provides the key path to the count field (`dayCount`, `weekCount`,
-    /// or `monthCount`) based on the activity period (`daily`, `weekly`, or `monthly`).
-    ///
     var countField: ReferenceWritableKeyPath<UserActivity, Int32> {
         switch self {
         case .daily:
@@ -49,11 +37,7 @@ enum ActivityPeriod: String, Identifiable, CaseIterable {
     }
 }
 
-// MARK: - Spread
-
 extension ActivityPeriod {
-
-    /// This property is used to determine the date range for creating or fetching `UserActivity` records.
     var spreadComponent: Calendar.Component {
         switch self {
         case .daily:
@@ -66,27 +50,16 @@ extension ActivityPeriod {
     }
 }
 
-// MARK: - ChartCompatible
-
-/// A protocol that defines properties and methods for chart compatibility.
-///
 extension ActivityPeriod: ChartCompatible {
-
-    /// A computed property that returns the range of dates for the activity period.
-    ///
     var range: Range<Date> {
         let today = Calendar(identifier: .iso8601).startOfDay(for: Date())
         return today.adding(rangeComponent, value: -1)..<today
     }
 
-    /// A computed property that returns the appropriate range component for the activity period.
-    ///
     var rangeComponent: Calendar.Component {
         return .month
     }
 
-    /// A computed property that returns the appropriate point component for the activity period.
-    ///
     var pointComponent: Calendar.Component {
         return .day
     }
