+++ Sources/Scout/Core/Matrix/PeriodCell.swift
