+++ Sources/Scout/Core/Monitor/IDs.swift
@@ -8,14 +8,7 @@
 import CoreData
 import Foundation
 
-/// An enumeration that encapsulates various identifiers used within the Scout application.
-/// This enumeration is intended to provide a centralized location for managing and accessing
-/// different types of IDs, ensuring consistency and reducing the risk of hardcoding values
-/// throughout the codebase.
-///
 enum IDs {
-
-    /// A static computed property that returns an optional UUID representing the session ID.
     static var session: UUID? {
         let context = persistentContainer.viewContext
         let request: NSFetchRequest<SessionObject> = SessionObject.fetchRequest()
@@ -26,12 +19,8 @@ enum IDs {
         return session?.sessionID
     }
 
-    /// A static constant that generates a new universally unique identifier (UUID) for a launch event.
     static let launch = UUID()
 
-    /// A static constant representing a unique identifier for a user.
-    /// The UUID is generated lazily when first accessed.
-    ///
     static let user: UUID = {
         let userKey = "scout_log_user_id"
 
@@ -45,23 +34,16 @@ enum IDs {
     }()
 }
 
-// MARK: - UUID
-
 extension UserDefaults {
-
-    /// Sets a UUID value in the user defaults for the specified key.
     fileprivate func set(_ value: UUID, forKey key: String) {
         set(value.uuidString, forKey: key)
     }
 
-    /// Retrieves a UUID value from the user defaults for the specified key.
     fileprivate func uuid(forKey key: String) -> UUID? {
         guard let string = string(forKey: key) else { return nil }
         return UUID(uuidString: string)
     }
 }
-
-// MARK: - Default Values
 
 extension SessionObject {
     public override func awakeFromInsert() {
