+++ Sources/Scout/Core/Monitor/NotificationListener.Activity.swift
