+++ Sources/Scout/Core/Monitor/SessionObject+Monitor.swift
