+++ Sources/Scout/Core/Monitor/UserActivity+Monitor.swift
