+++ Sources/Scout/Core/Storage/Persistence.swift
