+++ Sources/Scout/Core/Sync/CKPersistable.swift
