+++ Sources/Scout/Core/Telemetry/LogMetrics.swift
