+++ Sources/Scout/Core/Telemetry/Telemetry.Export.swift
