+++ Sources/Scout/Core/Telemetry/Telemetry.swift
