+++ Sources/Scout/Core/Telemetry/TelemetryFactory.swift
