+++ Sources/Scout/Core/Telemetry/TelemetryHandler.Idle.swift
