+++ Sources/Scout/Core/Telemetry/TelemetryHandler.swift
